fix: refactor comment service
