--- src/app/api/activity-logs/[id]/route.ts
@@ -1,4 +1,0 @@
-import { withErrorHandler } from '../../core/utils/withErrorHandler'
-import { getActivityWithComment } from '../activity.controller'
-
-export const GET = withErrorHandler(getActivityWithComment)
