+++ src/app/api/activity-logs/const.ts
@@ -2,9 +2,11 @@ import { ActivityType } from '@prisma/client'
 import { TaskCreatedSchema } from '@api/activity-logs/schemas/TaskCreatedSchema'
 import { TaskAssignedSchema } from '@api/activity-logs/schemas/TaskAssignedSchema'
 import { WorkflowStateUpdatedSchema } from '@api/activity-logs/schemas/WorkflowStateUpdatedSchema'
+import { CommentAddedSchema } from '@api/activity-logs/schemas/CommentAddedSchema'
 
 export const SchemaByActivityType = {
   [ActivityType.TASK_CREATED]: TaskCreatedSchema,
   [ActivityType.TASK_ASSIGNED]: TaskAssignedSchema,
   [ActivityType.WORKFLOW_STATE_UPDATED]: WorkflowStateUpdatedSchema,
+  [ActivityType.COMMENT_ADDED]: CommentAddedSchema,
 }
