+++ src/app/api/activity-logs/schemas/CommentAddedSchema.ts
@@ -1,0 +1,7 @@
+import { z } from 'zod'
+
+export const CommentAddedSchema = z.object({
+  id: z.string().uuid(),
+  content: z.string(),
+  initiatorId: z.string().uuid(),
+})
